[Bug]: DBAL Deprecations

+++ src/Controller/AdminController.php
@@ -35,13 +35,17 @@ class AdminController extends \Pimcore\Bundle\AdminBundle\Controller\AdminContro
         if ($request->get('data')) {
             if ($request->get('xaction') == 'destroy') {
                 $id = json_decode($request->get('data'), true);
-                $def = FavoriteOutputDefinition::getById($id['id']);
-                if (!empty($def)) {
-                    $def->delete();
+                $idValue = $id['id'] ?? '';
+                if(!empty($idValue)) {
+                    $def = FavoriteOutputDefinition::getById($idValue);
+                    if (!empty($def)) {
+                        $def->delete();
 
-                    return $this->adminJson(['data' => [], 'success' => true]);
-                } else {
-                    throw new \Exception('OutputDefinition with id ' . $id . ' not found.');
+                        return $this->adminJson(['data' => [], 'success' => true]);
+                    }
+                }
+                if(!$def) {
+                    throw new \Exception('OutputDefinition with id ' . $idValue . ' not found.');
                 }
             } elseif ($request->get('xaction') == 'update') {
                 $data = json_decode($request->get('data'), true);
